Find installed TNG version.
Make it possible to use external TNG or compile the version
shipped with GROMACS (default).
Bumped the version of the TNG library. Corresponds to commit
c1e87072beafaf5a22e7c31c1263f7571fdfdfec in the TNG repository.

+++ src/gromacs/fileio/tngio_for_tools.cpp
@@ -427,7 +427,7 @@ gmx_bool gmx_read_next_tng_frame(tng_trajectory_t            input,
     void                   *values        = NULL;
     double                  frameTime     = -1.0;
     int                     size, blockDependency;
-    float                   prec;
+    double                  prec;
     const int               defaultNumIds = 5;
     static gmx_int64_t      fallbackRequestedIds[defaultNumIds] =
     {
@@ -778,7 +778,7 @@ gmx_bool gmx_get_tng_data_next_frame_of_block_type(tng_trajectory_t     input,
     gmx_int64_t         codecId;
     int                 blockDependency;
     void               *data = 0;
-    float               localPrec;
+    double              localPrec;
 
     stat = tng_data_block_name_get(input, blockId, name, maxLen);
     if (stat != TNG_SUCCESS)
